ADD: Restructure repo, split up files
