+++ src/data/filters/__init__.py
@@ -1,0 +1,2 @@
+from .general_filters import *
+from .audio_filters import *
